Added windows implementation of aws_backtrace_print
